rename(apps/app): Tab 폴더 BottomTab으로 네이밍 변경
