+++ apps/app/src/screens/BottomTab/CreateDiaryScreen.tsx
